GH-1806: Move IBuildRequestFactory to server.build
